requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
aiosqlite==0.20.0
python-multipart==0.0.9
pydantic==2.7.1
PyMuPDF==1.24.1
pillow==10.3.0 
openai==1.30.1
httpx==0.27.2
python-dotenv==1.0.1
apscheduler==3.11.0